backend/app/requirements.txt
aiohttp==3.8.3
aioresponses==0.7.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.5.2
astroid==2.4.2
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.4
cryptography==38.0.4
debugpy==1.5.1
decorator==5.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dnsdumpster @ https://github.com/PaulSec/API-dnsdumpster.com/archive/master.zip
entrypoints==0.3
frozenlist==1.3.3
gevent==22.10.2
google==3.0.0
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.2
httpretty==1.1.4
httpx==0.23.1
humanize==4.4.0
idna==2.10
ipykernel==6.5.0
ipython==7.29.0
isort==5.6.4
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
lazy-object-proxy==1.4.3
lxml==4.6.3
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.3
nest-asyncio==1.5.1
numpy==1.20.1
opencv-python==4.5.1.48
parso==0.8.2
pickleshare==0.7.5
plotly==5.4.0
pluginbase==1.0.1
prompt-toolkit==3.0.22
proxy-db==0.3.1
pycparser==2.21
Pygments==2.10.0
pylint==2.6.0
PyQt6==6.1.1
PyQt6-Qt6==6.1.3
PyQt6-sip==13.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.7
python-Wappalyzer==0.3.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
PyYAML==5.3.1
pyzmq==22.3.0
requests==2.25.0
requests-mock==1.10.0
rfc3986==1.5.0
shiboken2==5.15.2
shiboken6==6.1.3
six==1.15.0
sniffio==1.3.0
soupsieve==2.0.1
SQLAlchemy==1.4.44
sqlparse==0.4.3
tenacity==8.0.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.2
waybackpy==3.0.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
